% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_integrity.R
\name{check_data_integrity}
\alias{check_data_integrity}
\title{Check data integrity}
\usage{
check_data_integrity(dataset)
}
\arguments{
\item{dataset}{Dataset to check the integrity}
}
\value{
Side effect: Display message in the command to say if the data integrity is OK or not
}
\description{
Check data integrity
}
\examples{
datafile <- system.file("nyc_squirrels_sample.csv", package = "checkdataanme")
nyc_squirrels <- read.csv(datafile)
check_data_integrity(nyc_squirrels)
}
