% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdataanme-package.R
\docType{package}
\name{checkdataanme-package}
\alias{checkdataanme}
\alias{checkdataanme-package}
\title{checkdataanme: First Fusen Full Package}
\description{
Create a package following the exercice description on slide 372.
}
\author{
\strong{Maintainer}: Aurélien Macé \email{aurelien.mace@finddx.org}

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
